requirements.txt
autopep8==1.6.0
krovak05==0.0.9
pycodestyle==2.8.0
pynmea2==1.18.0
pyserial==3.5
toml==0.10.2
